fix: リロードするたびに新しいユーザーが作成される不具合を修正
