+++ server/api/me/player/controller.ts
@@ -1,0 +1,7 @@
+import { defineController } from './$relay';
+
+export default defineController(() => ({
+  get: ({ player }) => {
+    return { status: 200, body: player };
+  },
+}));
