+++ server/api/me/player/index.ts
@@ -1,0 +1,7 @@
+import type { PlayerModel } from '$/commonTypesWithClient/models';
+
+export type Methods = {
+  get: {
+    resBody: PlayerModel;
+  };
+};
